+++ modules/tinymce/src/themes/silver/main/ts/ui/throbber/Throbber.ts
@@ -5,18 +5,19 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { AlloyComponent, AlloySpec, Behaviour, DomFactory, Focusing, Keying, Replacing } from '@ephox/alloy';
-import { Cell, Optional, Type } from '@ephox/katamari';
-import { Attribute, Css } from '@ephox/sugar';
+import { AlloyComponent, AlloySpec, Behaviour, Blocking, Composing, DomFactory, Replacing } from '@ephox/alloy';
+import { Arr, Cell, Optional, Type } from '@ephox/katamari';
+import { Attribute, Css, Focus, SugarElement } from '@ephox/sugar';
 import Editor from 'tinymce/core/api/Editor';
 import Delay from 'tinymce/core/api/util/Delay';
 import { UiFactoryBackstageProviders, UiFactoryBackstageShared } from '../../backstage/Backstage';
 
-const renderSpinner = (providerBackstage: UiFactoryBackstageProviders): AlloySpec => ({
+const getBusySpec = (providerBackstage: UiFactoryBackstageProviders) => (_root: AlloyComponent, _behaviours: Behaviour.AlloyBehaviourRecord): AlloySpec => ({
   dom: {
     tag: 'div',
     attributes: {
-      'aria-label': providerBackstage.translate('Loading...')
+      'aria-label': providerBackstage.translate('Loading...'),
+      'tabindex': '0'
     },
     classes: [ 'tox-throbber__busy-spinner' ]
   },
@@ -25,27 +26,56 @@ const renderSpinner = (providerBackstage: UiFactoryBackstageProviders): AlloySpe
       dom: DomFactory.fromHtml('<div class="tox-spinner"><div></div><div></div><div></div></div>')
     }
   ],
-  behaviours: Behaviour.derive([
-    // Trap the "Tab" key and don't let it escape.
-    Keying.config({
-      mode: 'special',
-      onTab: () => Optional.some(true),
-      onShiftTab: () => Optional.some(true)
-    }),
-    Focusing.config({ })
-  ])
 });
 
-const toggleThrobber = (comp: AlloyComponent, state: boolean, providerBackstage: UiFactoryBackstageProviders) => {
+const focusBusyComponent = (throbber: AlloyComponent): void =>
+  Composing.getCurrent(throbber).each((comp) => Focus.focus(comp.element));
+
+// When the throbber is enabled, prevent the iframe from being part of the sequential keyboard navigation when Tabbing
+// TODO: TINY-7500 Only works for iframe mode at this stage
+const toggleEditorTabIndex = (editor: Editor, state: boolean) => {
+  const tabIndexAttr = 'tabindex';
+  const dataTabIndexAttr = `data-mce-${tabIndexAttr}`;
+  Optional.from(editor.iframeElement)
+    .map(SugarElement.fromDom)
+    .each((iframe) => {
+      if (state) {
+        Attribute.getOpt(iframe, tabIndexAttr).each((tabIndex) => Attribute.set(iframe, dataTabIndexAttr, tabIndex));
+        Attribute.set(iframe, tabIndexAttr, -1);
+      } else {
+        Attribute.remove(iframe, tabIndexAttr);
+        Attribute.getOpt(iframe, dataTabIndexAttr).each((tabIndex) => {
+          Attribute.set(iframe, tabIndexAttr, tabIndex);
+          Attribute.remove(iframe, dataTabIndexAttr);
+        });
+      }
+    });
+};
+
+/*
+* If the throbber has been toggled on, only focus the throbber if the editor had focus as we don't to steal focus if it is on an input or dialog
+* If the throbber has been toggled off, only put focus back on the editor if the throbber had focus.
+* The next logical focus transition from the throbber is to put it back on the editor
+*/
+const toggleThrobber = (editor: Editor, comp: AlloyComponent, state: boolean, providerBackstage: UiFactoryBackstageProviders) => {
   const element = comp.element;
-  if (state === true) {
-    Replacing.set(comp, [ renderSpinner(providerBackstage) ]);
+  toggleEditorTabIndex(editor, state);
+  if (state) {
+    Blocking.block(comp, getBusySpec(providerBackstage));
     Css.remove(element, 'display');
     Attribute.remove(element, 'aria-hidden');
+    if (editor.hasFocus()) {
+      focusBusyComponent(comp);
+    }
   } else {
-    Replacing.set(comp, [ ]);
+    // Get the focus of the busy component before it is removed from the DOM
+    const throbberFocus = Composing.getCurrent(comp).exists((busyComp) => Focus.hasFocus(busyComp.element));
+    Blocking.unblock(comp);
     Css.set(element, 'display', 'none');
     Attribute.set(element, 'aria-hidden', 'true');
+    if (throbberFocus) {
+      editor.focus();
+    }
   }
 };
 
@@ -62,7 +92,13 @@ const renderThrobber = (spec): AlloySpec => ({
     }
   },
   behaviours: Behaviour.derive([
-    Replacing.config({})
+    Replacing.config({}),
+    Blocking.config({
+      focus: false
+    }),
+    Composing.config({
+      find: (comp) => Arr.head(comp.components())
+    })
   ]),
   components: [ ]
 });
@@ -71,10 +107,33 @@ const setup = (editor: Editor, lazyThrobber: () => AlloyComponent, sharedBacksta
   const throbberState = Cell<boolean>(false);
   const timer = Cell<Optional<number>>(Optional.none());
 
+  const stealFocus = (e) => {
+    if (throbberState.get()) {
+      e.preventDefault();
+      focusBusyComponent(lazyThrobber());
+      editor.editorManager.setActive(editor);
+    }
+  };
+
+  // TODO: TINY-7500 Only worrying about iframe mode at this stage since inline mode has a number of other issues
+  if (!editor.inline) {
+    editor.on('PreInit', () => {
+      // Cover focus when when the editor is focused natively
+      editor.dom.bind(editor.getWin(), 'focusin', stealFocus);
+      // Cover stealing focus when editor.focus() is called
+      editor.on('BeforeExecCommand', (e) => {
+        // If skipFocus is specified as true in the command, don't focus the Throbber
+        if (e.command.toLowerCase() === 'mcefocus' && e.value !== true) {
+          stealFocus(e);
+        }
+      });
+    });
+  }
+
   const toggle = (state: boolean) => {
     if (state !== throbberState.get()) {
-      toggleThrobber(lazyThrobber(), state, sharedBackstage.providers);
       throbberState.set(state);
+      toggleThrobber(editor, lazyThrobber(), state, sharedBackstage.providers);
       editor.fire('AfterProgressState', { state });
     }
   };